refactor(harness): extract skills engine with config injection
